Factor packaging into subfolder of tools
