+++ tools/packaging/dmg_settings.json
